requirements.txt
streamlit
pandas
openai
python-dotenv
PyPDF2
xlsxwriter
